functions/feedback-receiver/requirements.txt
flask==2.0.1
google-cloud-pubsub==2.12.0
gunicorn==20.1.0
